+++ src/App.js
@@ -1,11 +1,8 @@
 
 import './App.css';
 import {Route, Routes, NavLink } from 'react-router-dom';
-import Home from './Component/Home';
-import Bezawit from './Component/Bezawit';
+import Bezawit from './Components/Bezawit';
 
-
-import {NavLink,Routes,Route} from 'react-router-dom';
 //import Bezawit from './Components/Bezawit'
 //import Kavitha from './Components/Kavitha'
 import Poovai from './Components/Poovai'
@@ -13,7 +10,7 @@ import Home from './Components/Home'
 
 
 
-import Imane from '../Components/Imane'
+import Imane from './Components/Imane'
 
 
 
@@ -26,6 +23,7 @@ function App() {
 
             <NavLink to={"/"}>Home</NavLink>
             <NavLink to={"/poovai"}>Poovai</NavLink>
+            <NavLink to={"/bezawit"}>Bezawit</NavLink>
           
           </nav>
           
@@ -34,6 +32,7 @@ function App() {
           
           <Route path="/" element={<Home/>}/>
           <Route path="/poovai" element={<Poovai/>}/>
+          <Route path="/bezawit" element={<Bezawit/>}/>
           
           </Routes>
