--- src/Component/Bezawit.js
@@ -1,31 +1,0 @@
-function Bezawit() {
-    return(
-        <div>
-        <div className='img-bezawit'>
-                {/* <img src=""/> */}
-                <div className="bezawit-info">
-                    <h1>Bezawit Gebru</h1>
-                <h2>Entry Level Full Stack Web Developer</h2>
-                <p>Student at Perscholas</p>
-                <p>Boston, MA</p>
-                <ul>
-                    <li>Html</li>
-                    <li>CSS3</li>
-                    <li>React</li>
-                    
-                    </ul>
-                </div>
-                
-
-                </div>
-                <h1>Projects</h1>
-                <div className="project-app">
-                    <p>University Search App</p>
-                </div>
-                
-        </div>
-
-    )
-}
-
-export default Bezawit
