--- src/Component/Home.js
@@ -1,7 +1,0 @@
-function Home(){
-    return(
-        <div>Home</div>
-    )
-}
-
-export default Home
