+++ src/Components/Bezawit.js
@@ -1,0 +1,39 @@
+function Bezawit() {
+    return(
+        <div>
+        <div className='img-bezawit'>
+                {/* <img src=""/> */}
+                <div className="bezawit-info">
+                    <h1>Bezawit Gebru</h1>
+                <h2>Junior Full Stack Web Developer</h2>
+               
+               
+                <h3>Skills</h3>
+                
+                <div className="row">
+                    
+                    <p>Html5</p>
+                    <p>CSS3</p>
+                    <p>Javascript</p>
+                    <p>React</p>
+                    <p>Mongo</p>
+                    <p>Express</p>
+                    <p>Github</p>
+                    
+                    </div>
+                    <p>Boston, MA</p>
+                </div>
+                
+
+                </div>
+                {/* <h1>Projects</h1> */}
+                {/* <div className="project-app">
+                    <p>University Search App</p>
+                </div> */}
+                
+        </div>
+
+    )
+}
+
+export default Bezawit
